\name{internals}

\alias{ADEgORtrellis}
\alias{ADEgORtrellis-class}
\alias{ADEgORADEgSORtrellis}
\alias{ADEgORADEgSORtrellis-class}

\alias{axis.L}
\alias{setlatticecall}
\alias{setlatticecall-methods}
\alias{setvalueskey}

\alias{printSuperpose}
\alias{printSuperpose-methods}

\alias{as.raster.pixmapGrey}
\alias{as.raster.pixmapRGB}
\alias{panel.symbols.grid}

\title{For internal use only}

\description{
  Internal classes, methods and functions. Not for users.
}

\author{Alice Julien-Laferriere, Aurelie Siberchicot \email{aurelie.siberchicot@univ-lyon1.fr} and Stephane Dray
}

\keyword{internal}